% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks-advanced.R
\name{callback_context}
\alias{callback_context}
\title{TODO document}
\usage{
callback_context()
}
\description{
TODO document
}
\examples{
dash_app() \%>\%
  set_layout(
    button('Button 1', id='btn1'),
    button('Button 2', id='btn2'),
    button('Button 3', id='btn3'),
    div(id='container')
  ) \%>\%
  add_callback(
    "container",
    list(
      input("btn1", "n_clicks"),
      input("btn2", "n_clicks"),
      input("btn3", "n_clicks")
    ),
    function(input) {
      ctx <- callback_context()
      prevent_update(is.null(ctx))
      sprintf("Triggered: \%s, btn1: \%s, btn2: \%s, btn3: \%s",
              ctx$triggered$prop_id, input$btn1, input$btn2, input$btn3)
    }
  ) \%>\%
  run_app()
}
