% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{dash_tag}
\alias{dash_tag}
\title{Document that className and htmlFor is used instead of class and for
Document that style is given as a list rather than a string}
\usage{
dash_tag(tag_name, content = list(), n_clicks = NULL)
}
\description{
Document that className and htmlFor is used instead of class and for
Document that style is given as a list rather than a string
}
