% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash_app.R
\name{print.dash2}
\alias{print.dash2}
\title{Run a Dash app when explicitly printed to the console}
\usage{
\method{print}{dash2}(x, ...)
}
\description{
Run a Dash app when explicitly printed to the console
}
\keyword{internal}
