% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback.R
\name{add_callback}
\alias{add_callback}
\title{Add a callback to a Dash app}
\usage{
add_callback(app, outputs, params, callback)
}
\arguments{
\item{app}{A dash application created with \code{\link[=dash_app]{dash_app()}}.
TODO document
TODO make sure clientside callbacks work https://dashr.plotly.com/clientside-callbacks
TODO consider making property=value and property=children the default for input/output (looking through many example apps, it's very common)
TODO consider allowing user to pass in a string instead of a input/output object}
}
\description{
Add a callback to a Dash app
}
